[ANDROID-110] feat : 로그아웃 로직에도 Auth 헤더를 넣도록 설정
