+++ core/data/src/main/java/see/day/network/AuthService.kt
@@ -3,7 +3,6 @@ package see.day.network
 import retrofit2.http.Body
 import retrofit2.http.POST
 import see.day.network.dto.CommonResponse
-import see.day.network.dto.auth.LogoutRequest
 import see.day.network.dto.auth.RefreshTokenRequest
 import see.day.network.dto.auth.RefreshTokenResponse
 import see.day.network.dto.login.LoginRequest
@@ -16,7 +15,4 @@ interface AuthService {
 
     @POST("api/auth/refresh")
     suspend fun refresh(@Body refreshTokenRequest: RefreshTokenRequest): CommonResponse<RefreshTokenResponse>
-
-    @POST("api/auth/logout")
-    suspend fun logout(@Body logoutRequest: LogoutRequest) : CommonResponse<Unit>
 }
