+++ core/data/src/main/java/see/day/network/UserService.kt
@@ -7,6 +7,7 @@ import retrofit2.http.POST
 import retrofit2.http.PUT
 import see.day.network.dto.CommonResponse
 import see.day.network.dto.auth.DeleteUserRequest
+import see.day.network.dto.auth.LogoutRequest
 import see.day.network.dto.user.FullUserResponse
 import see.day.network.dto.user.OnboardingCompleteRequest
 import see.day.network.dto.user.UserProfileChangedInputRequest
@@ -24,4 +25,8 @@ interface UserService {
 
     @PUT("api/users/profile")
     suspend fun updateUserProfile(@Body userProfileChangedInputRequest: UserProfileChangedInputRequest) : CommonResponse<FullUserResponse>
+
+    // accessToken이 필요하므로 UserService로 옮김
+    @POST("api/auth/logout")
+    suspend fun logout(@Body logoutRequest: LogoutRequest) : CommonResponse<Unit>
 }
