+++ core/data/src/main/java/see/day/repository/LoginRepositoryImpl.kt
@@ -1,6 +1,5 @@
 package see.day.repository
 
-import android.util.Log
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -23,13 +22,12 @@ import see.day.model.navigation.AppStartState.HOME
 import see.day.model.navigation.AppStartState.LOGIN
 import see.day.model.navigation.AppStartState.ONBOARDING
 import see.day.network.AuthService
-import see.day.network.dto.auth.LogoutRequest
 import see.day.network.dto.auth.RefreshTokenRequest
 import see.day.utils.ErrorUtils.createResult
 
 class LoginRepositoryImpl @Inject constructor(
     private val dataSource: DataStoreDataSource,
-    @Auth private val authService: AuthService
+    @Auth private val authService: AuthService,
 ) : LoginRepository {
 
     override suspend fun login(socialLogin: SocialLogin): Result<AppStartState> {
@@ -46,22 +44,6 @@ class LoginRepositoryImpl @Inject constructor(
             } else {
                 ONBOARDING
             }
-        }
-    }
-
-    override suspend fun logout(allDevices: Boolean): Result<Unit> {
-        return createResult {
-            val refreshToken = dataSource.getRefreshToken().first()
-            if(refreshToken.isNullOrEmpty()) {
-                dataSource.clearData()
-                throw NoDataException()
-            }
-            val logoutRequest = LogoutRequest(refreshToken, allDevices)
-
-            authService.logout(logoutRequest = logoutRequest)
-            dataSource.clearData()
-        }.onFailure {
-            dataSource.clearData()
         }
     }
