--- core/data/src/test/java/see/day/data/api/authService/json/LogoutJson.kt
@@ -1,9 +1,0 @@
-package see.day.data.api.authService.json
-
-val logoutSuccessResponse = """
-    {
-    "statusCode": 200,
-    "code": "S200",
-    "message": "로그아웃되었습니다."
-    }
-""".trimIndent()
